quick fix for failed unit test caused by index repository initialisation

+++ src/test/java/uk/ac/shef/dcs/jate/app/BaseEmbeddedSolrTest.java
@@ -13,34 +13,43 @@ import org.apache.solr.common.SolrDocumentList;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.core.CoreContainer;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.*;
 import uk.ac.shef.dcs.jate.JATEException;
 import uk.ac.shef.dcs.jate.JATEProperties;
 
-// Java Code Examples for EmbeddedSolrServer
-
-// http://stackoverflow.com/questions/936723/testing-solr-via-embedded-server
-// https://wiki.searchtechnologies.com/index.php/Unit_Testing_with_Embedded_Solr
-// http://www.programcreek.com/java-api-examples/index.php?api=org.apache.solr.client.solrj.embedded.EmbeddedSolrServer
+/**
+ * Abstracted class to support embedded solr based unit test
+ *
+ * for more examples about how to use solr Embedded Solr server,
+ *
+ * @see <a href="http://www.programcreek.com/java-api-examples/index.php?api=org.apache.solr.client.solrj.embedded.EmbeddedSolrServer">
+ *     Java Code Examples for org.apache.solr.client.solrj.embedded.EmbeddedSolrServer</a>
+ *
+ * @see <a href="https://wiki.searchtechnologies.com/index.php/Unit_Testing_with_Embedded_Solr">
+ *     Unit Testing with Embedded Solr</a>
+ *
+ */
 public abstract class BaseEmbeddedSolrTest {
     private static Logger LOG = Logger.getLogger(BaseEmbeddedSolrTest.class.getName());
 
-    String workingDir = System.getProperty("user.dir");
-    String solrCoreName = "testCore";
-    String solrHome = workingDir + "\\testdata\\solr-testbed\\";
+    static String workingDir = System.getProperty("user.dir");
+    static String solrCoreName = "testCore";
+    static String solrHome = workingDir + "\\testdata\\solr-testbed\\";
 
     EmbeddedSolrServer server;
 
-    @Before
-    public void setup() throws Exception {
+    @BeforeClass
+    public static void setupClass() throws Exception {
         try {
-        cleanIndexDirectory(solrHome, solrCoreName);
+            cleanIndexDirectory(solrHome, solrCoreName);
         } catch (IOException ioe) {
-            throw new JATEException("Unable to delete index data. Please clean index directory [testdata\\solr-testbed\\testCore\\data] manually!");
+            throw new JATEException("Unable to delete index data. Please clean index directory " +
+                    "[testdata\\solr-testbed\\testCore\\data] manually!");
         }
+    }
 
+    //@Before
+    public void setup() throws Exception {
         CoreContainer testBedContainer = new CoreContainer(solrHome);
         testBedContainer.load();
 
@@ -50,11 +59,11 @@ public abstract class BaseEmbeddedSolrTest {
     /**
      * create and add new document with necessary fields
      *  see also @code{uk.ac.shef.dcs.jate.indexing.IndexingHandler}
-     * @param docId
-     * @param docTitle
-     * @param text
-     * @param jateProperties
-     * @param commit
+     * @param docId, document id value for default solr field 'id'
+     * @param docTitle, document title for default solr field 'title_s'
+     * @param text, document content for default solr field 'text' where term will be extracted
+     * @param jateProperties, JATE properties for various run-time parameters
+     * @param commit, boolean value to determine whether the new document will be committed
      * @throws IOException
      * @throws SolrServerException
      * @throws JATEException
@@ -76,37 +85,12 @@ public abstract class BaseEmbeddedSolrTest {
         }
     }
 
-//    @Test
-//    public void testDocInsertion() throws IOException, SolrServerException {
-//
-//        addNewDoc("doc-1", "Test Document 1", "Hello solr!", true);
-//
-//        ModifiableSolrParams params = new ModifiableSolrParams();
-//        params.set("q", "*:*");
-//
-//        try {
-//            QueryResponse qResp = server.query(params);
-//            SolrDocumentList docList = qResp.getResults();
-//
-//            assert (docList.getNumFound() == 1);
-//            /**
-//            cleanData();
-//            QueryResponse qResp2 = server.query(params);
-//            docList = qResp.getResults();
-//            System.out.println("Num docs after deletion: " + docList.getNumFound());
-//            assert (docList.getNumFound() == 0);
-//            **/
-//        } catch (SolrServerException e) {
-//            e.printStackTrace();
-//        } catch (IOException ioe) {
-//            ioe.printStackTrace();
-//        }
-//    }
-
-    public void cleanIndexDirectory(String solrHome, String coreName) throws IOException {
+    public static void cleanIndexDirectory(String solrHome, String coreName) throws IOException {
         File indexDir = new File(solrHome + "/" + coreName + "/data/index/");
         try {
-            FileUtils.cleanDirectory(indexDir);
+            if (indexDir.exists()) {
+                FileUtils.cleanDirectory(indexDir);
+            }
         } catch (IOException e) {
             LOG.error("Failed to clean index directory! Please do it manually!");
             throw e;
@@ -114,7 +98,6 @@ public abstract class BaseEmbeddedSolrTest {
     }
 
     public void cleanData() throws IOException, SolrServerException {
-        // LOG.log(Level.FINE, );
         LOG.info("clean all test data...");
         try {
             server.deleteByQuery("*:*");
@@ -125,39 +108,39 @@ public abstract class BaseEmbeddedSolrTest {
         }
     }
 
-    @After
+    //@After
     public void tearDown() throws IOException {
         if (server != null && server.getCoreContainer() != null) {
             LOG.info("shutting down core in :" + server.getCoreContainer().getCoreRootDirectory());
 
-//            try {
-//                cleanData();
-//            } catch (IOException e) {
-//                e.printStackTrace();
-//                LOG.error("Failed to clean test data. Please do it manually!");
-//            } catch (SolrServerException e) {
-//                e.printStackTrace();
-//                LOG.error("Failed to clean test data. Please do it manually!");
-//            }
             server.getCoreContainer().getCore(solrCoreName).close();
             server.getCoreContainer().shutdown();
             server.close();
 
             try {
                 // wait for server shutting down
-                Thread.sleep(6000);
+                Thread.sleep(10000);
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
-            try {
-                cleanIndexDirectory(solrHome, solrCoreName);
-            } catch (IOException ioe) {
-                LOG.warn("Unable to delete indice file. Please do it manually! ");
-            }
+//            try {
+//                cleanIndexDirectory(solrHome, solrCoreName);
+//            } catch (IOException ioe) {
+//                LOG.warn("Unable to delete indice file. Please do it manually! ");
+//            }
         } else {
-            // LOG.log(Level.SEVERE, "embedded server or core is not created and loaded successfully.");
             LOG.error("embedded server or core is not created and loaded successfully.");
         }
+    }
+
+    @AfterClass
+    public static void tearDownClass() throws JATEException {
+//        try {
+//            cleanIndexDirectory(solrHome, solrCoreName);
+//        } catch (IOException ioe) {
+//            throw new JATEException("Unable to delete index data. Please clean index directory " +
+//                    "[testdata\\solr-testbed\\testCore\\data] manually!");
+//        }
     }
 
 }
